ai-engine/requirements.txt
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.0
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
geoip2==4.7.0
ipaddress==1.0.23
schedule==1.2.0
pytest==7.4.0
pytest-cov==4.1.0